+++ src/Strategy/CommonClassesStrategy.php
@@ -47,6 +47,8 @@ use Zend\Diactoros\Request as ZendDiactorosRequest;
  * @internal
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ *
+ * Don't miss updating src/Composer/Plugin.php when adding a new supported class.
  */
 final class CommonClassesStrategy implements DiscoveryStrategy
 {