+++ src/Strategy/CommonPsr17ClassesStrategy.php
@@ -13,6 +13,8 @@ use Psr\Http\Message\UriFactoryInterface;
  * @internal
  *
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
+ *
+ * Don't miss updating src/Composer/Plugin.php when adding a new supported class.
  */
 final class CommonPsr17ClassesStrategy implements DiscoveryStrategy
 {